+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Change Log
 
+## v0.0.8
+- Fix bug when open multiple window on same file for soundevents
+- Rewrite HeroList Editor
+
+## v0.0.7
+- Change request events
+
+## v0.0.6
+- Support initialize custom_net_tables.txt
+- Support initialize hero list
+
 ## v0.0.5
 - Remove some buttons event for SoundEvents Editor